DBOPS-853 - Alters for credits and credit_transactions (#23457)
* DBOPS-853

* migration changes

* added unsigned for columns

* Update database/queries/queries.txt

+++ database/migrations/2016_07_30_123142_create_credits_table.php
@@ -31,11 +31,11 @@ class CreateCreditsTable extends Migration
             $table->char(Credits::PROMOTION_ID, Promotion\Entity::ID_LENGTH)
                   ->nullable();
 
-            $table->integer(Credits::VALUE);
+            $table->bigInteger(Credits::VALUE);
 
             $table->string(Credits::TYPE, 20);
 
-            $table->integer(Credits::USED)
+            $table->bigInteger(Credits::USED)
                   ->default(0);
 
             $table->integer(Credits::EXPIRED_AT)
